+++ src/css/index.css
@@ -4,7 +4,6 @@
 
 body {
   font-family: 'Pretendard', sans-serif;
-  background-color: #FCF4EF;
   max-width: 1024px; /* 기본 너비 설정 */
   margin: 0 auto;    /* 중앙 정렬 */
   padding: 0 1rem;  /* 기본 패딩 (필요에 따라 조정) */
